+++ crawler/src/Robots.h
@@ -1,6 +1,8 @@
 #ifndef CRAWLER_ROBOTS_H
 #define CRAWLER_ROBOTS_H
 
+#include "core/lru_cache.h"
+#include "core/optional.h"
 #include "http/RequestExecutor.h"
 #include "http/Response.h"
 #include "http/URL.h"
@@ -12,13 +14,13 @@
 #include <queue>
 #include <string>
 #include <string_view>
-#include <unordered_map>
 #include <utility>
 #include <vector>
 
 namespace mithril {
 
-constexpr long RobotsTxtCacheDurationSeconds = 4L * 60L * 60L;  // 4 hours
+constexpr long RobotsTxtCacheDurationSeconds = 4L * 60L * 60L;         // 4 hours
+constexpr long RobotsTxtCacheFailureDurationSeconds = 1L * 60L * 60L;  // 1 hour
 
 namespace internal {
 
@@ -37,6 +39,7 @@ private:
         std::vector<std::pair<std::string, Node>> fixedSegments;
         std::unique_ptr<Node> wildcardMatch;  // corresponds to a *
         std::unique_ptr<Node> emptyMatch;     // corresponds to a / with nothing before it
+        bool trailingWildcard{false};
         NodeType type{NodeType::NonTerminal};
         uint16_t patternLength{0};
     };
@@ -68,6 +71,7 @@ struct RobotLine {
 };
 
 struct RobotDirectives {
+    core::Optional<unsigned long> crawlDelay;
     std::vector<std::string> disallows;
     std::vector<std::string> allows;
 };
@@ -85,13 +89,9 @@ public:
      */
     RobotRules();
 
-    /**
-     * @brief Creates a RobotRules object that either allows or disallows all
-     * paths.
-     *
-     * @param disallowAll Whether to disallow all paths.
-     */
-    RobotRules(bool disallowAll);
+    static RobotRules AllowAll();
+
+    static RobotRules DisallowAll();
 
     /**
      * @brief Creates a RobotRules object from a list of disallowed prefixes and
@@ -99,8 +99,11 @@ public:
      *
      * @param disallowPrefixes Disallowed prefixes to filter by.
      * @param allowPrefixes Allowed prefixes to filter by.
+     * @param crawlDelay Crawl delay for host, if specified.
      */
-    RobotRules(const std::vector<std::string>& disallowPrefixes, const std::vector<std::string>& allowPrefixes);
+    RobotRules(const std::vector<std::string>& disallowPrefixes,
+               const std::vector<std::string>& allowPrefixes,
+               core::Optional<unsigned long> crawlDelay);
 
     /**
      * @brief Creates a RobotRules object from the contents of a robots.txt file.
@@ -120,15 +123,29 @@ public:
      */
     bool Allowed(std::string_view path) const;
 
+    /**
+     * @brief Gets the Crawl-Delay, if specified.
+     */
+    const core::Optional<unsigned long>& CrawlDelay() const;
+
 private:
+    /**
+     * @brief Creates a RobotRules object that either allows or disallows all
+     * paths.
+     *
+     * @param disallowAll Whether to disallow all paths.
+     */
+    RobotRules(bool disallowAll);
+
     std::unique_ptr<internal::RobotsTrie> trie_;
     bool disallowAll_;
+    core::Optional<unsigned long> crawlDelay_;
 };
 
 
 class RobotRulesCache {
 public:
-    RobotRulesCache(size_t maxInFlightRequests);
+    RobotRulesCache(size_t maxInFlightRequests, size_t cacheSize);
 
     /**
      * @brief Gets the ruleset associated with the canonical host, or queues up
@@ -160,8 +177,9 @@ private:
     struct RobotCacheEntry {
         RobotRules rules;
         long expiresAt{0L};
-        bool valid{false};
     };
+
+    void FillFromQueue();
 
     /**
      * @brief Enqueues a host for robots.txt fetching.
@@ -185,8 +203,7 @@ private:
 
     size_t maxInFlightRequests_;
 
-    // TODO: Use an LRU cache
-    std::unordered_map<std::string, RobotCacheEntry> cache_;
+    core::LRUCache<std::string, RobotCacheEntry> cache_;
     std::queue<http::CanonicalHost> queuedFetches_;
     http::RequestExecutor executor_;